fix: 합성 컴포넌트 사용하는 방식으로 수정
